support custom CA bundle for AWS API
When a cluster is installed in a AWS C2S region, access to the AWS
API requires a custom CA bundle for trust. The custom CA bundle
is read from the "ca-bundle.pem" key of the kube-cloud-config
ConfigMap in the openshift-config-managed namespace.

To get the EFS provisioner pod to use the custom CA bundle, the
operator will create a ConfigMap in the namespace of the provisioner
that contains a copy of the CA bundle read from the kube-cloud-config
ConfigMap. The operator will adjust the deployment for the EFS
provisioner so that it mounts the custom CA bundle and uses the
AWS_CA_BUNDLE environment variable to point to the CA bundle. The
AWS SDK will use the AWS_CA_BUNDLE environment variable directly.

https://issues.redhat.com/browse/CORS-1584